platform/requirements.txt
# Project requirements (install w/ `python -m pip install -r requirements.txt`)
fastapi
uvicorn
tortoise-orm[asyncpg]
